fix removing the dependent entities

+++ src/main/java/com/estateagency/service/BuyerServiceImpl.java
@@ -1,7 +1,10 @@
 package com.estateagency.service;
 
 import com.estateagency.dao.BuyerDao;
+import com.estateagency.dao.ContractDao;
 import com.estateagency.model.Buyer;
+import com.estateagency.model.Contract;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -10,6 +13,9 @@ import java.util.List;
 @Service
 public class BuyerServiceImpl implements BuyerService {
     private BuyerDao buyerDao;
+
+    @Autowired
+    private ContractDao contractDao;
 
     public void setBuyerDao(BuyerDao buyerDao) {
         this.buyerDao = buyerDao;
@@ -29,8 +35,13 @@ public class BuyerServiceImpl implements BuyerService {
 
     @Override
     @Transactional
-    public void removeBuyer(int id) {
+    public boolean removeBuyer(int id) {
+        List<Contract> contracts = this.contractDao.getContractByFlat(id);
+        if (contracts.size() > 0) {
+            return false;
+        }
         this.buyerDao.removeBuyer(id);
+        return true;
     }
 
     @Override
