+++ src/main/java/com/estateagency/service/SellerServiceImpl.java
@@ -1,7 +1,10 @@
 package com.estateagency.service;
 
+import com.estateagency.dao.ContractDao;
+import com.estateagency.model.Contract;
 import com.estateagency.model.Seller;
 import com.estateagency.dao.SellerDao;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -10,6 +13,9 @@ import java.util.List;
 @Service
 public class SellerServiceImpl implements SellerService {
     private SellerDao sellerDao;
+
+    @Autowired
+    private ContractDao contractDao;
 
     public void setSellerDao(SellerDao sellerDao) {
         this.sellerDao = sellerDao;
@@ -29,8 +35,13 @@ public class SellerServiceImpl implements SellerService {
 
     @Override
     @Transactional
-    public void removeSeller(int id) {
+    public boolean removeSeller(int id) {
+        List<Contract> contracts = this.contractDao.getContractByFlat(id);
+        if (contracts.size() > 0) {
+            return false;
+        }
         this.sellerDao.removeSeller(id);
+        return true;
     }
 
     @Override
